backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-decouple==3.8
django-filter==23.2
django-crispy-forms==2.0
crispy-bootstrap5==0.7